requirements.txt
# VK Bot Dependencies

# Async VK API client
vkbottle==4.8.2

# HTTP requests
aiohttp==3.9.0

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
